Finish the whole release date countdown with the whole functionality.
Finished all the components that are involved in the container ReleaseDate.js
so that the whole container is ready. Main part in here is finishing the functionality
of the conuntdown time via /src/functions/ReleaseDate/findCurrentDifference.js
with help og moment.js. The interval is set in the /src/components/Counter.js
component, that fincd the current difference between the presence and the
release date, which is set in the function as a constant for
2022/10/18 18:00:00. Every time the interval runs (each second) the redux state
is updated with also correct form of string related to the actual value.
The strings are loaded from /src/conatstants/strings.js.