Add more rparts to the sign up page
Crated the rest of the sign up page. For now, the inputs are
real-time-validated and updated in the global Redux state. The prgress
bars (pwd, whole form) update in real time according to the state. On
finishing the form, there is no fuctionality so far. In the next commit,
we will try to add image cropper for the profile picture.